Former-commit-id: eb8db1fba7bf1a4db37fdfe19e20056f2c8f2d88

+++ backend/api/models/review.py
@@ -8,7 +8,7 @@ class Review(models.Model):
     # id = models.IntegerField(primary_key=True)
     store = models.ForeignKey(Store, on_delete=models.CASCADE)
     user = models.ForeignKey(User, on_delete=models.CASCADE)
-    score = models.IntegerField(null=True)
+    score = models.FloatField(null=True)
     content = models.TextField(null=True)
     reg_time = models.DateTimeField(auto_now_add=True, null=True)
